Draft wrappers/options
Options can be specified to define some more behaviour on validations.
As of now, there are three options:
1. if: validation is only performed if given proc/lambda returns true
2. unless: validation is only performed if given proc/lamda returns
false
3. allow_nil: validation is only performed if the attribute is not nil

Usage:
class UserValidator < Allowing::Validator
  validates :name, presence: true, unless: proc { |user| user.admin? }
end

Internally, each option is converted into a wrapper which contains a
rule (e.g. the proc) and a validation. When validate is called on the
wrapper, it is up to the wrapper to decide whether to pass it on to the
underlying validation or not