Disable trivial accessors in rubocop
The reason for this disagreement with Rubocop is as follows: First, a reader
can be defined explicitly for lazy initialization. I prefer to see a
reader and writer together and prefer to define all attr_* at the start
of the class definition. It is impossible to do this if a reader is
defined explicitly and the writer is not.

Second, there doesn't seem to be a way to define attr_* on a class
without opening the eigenclass. In this case, I prefer to write the
writer explicitly.