Add ValidationsBuilder with responsibility of manager
The ValidationsManager had to pass around a lot of objects to create the
correct validations. The ValidationsManager has been replaced with the
ValidationsBuilder which is responsible for building validations with
the attributes and rules of a single '#validates' statement. The
ValidationsManager did all validations for a single group. The new method
allows for more efficient use of instance variables instead of passing
variables via arguments.

Also the ValidationBuilder has been renamed to
AttributeValidationBuilder, to avoid confusion with the
ValidationsBuilder.