Allow multiple strategies for getting attribute
When accessing an attribute of the object being validated, there now are
multiple strategies for doing so. The strategy can be either passed to
the validator on initialization or it can be in the default_options
method of the validator. The key for this option is attribute_strategy.
The strategy can be a proc that takes the object being validated and the
attribute as arguments and it should return the value for that
attribute.

In addition to a proc, one of the following symbols can also be used:
* :send - The default strategy which sends the attribute to the object
* :symbol_key - Tries to get the attribute using the #[] method, using a
  symbol key
* :string_key - Tries to get the attribute using the #[] method, using a
  string key